place Tag & Card in subdirectory with styles.js
add TagContainer to Card’s styles.js
