+++ components/blog/Tag/Tag.js
@@ -1,0 +1,5 @@
+import S from './styles';
+
+export const Tag = ({ text }) => {
+  return <S.TagLink href={`/blog/category/${text}`}>{text}</S.TagLink>;
+};
