+++ components/containers/Card/Card.js
@@ -1,7 +1,8 @@
 import Image from 'next/image';
 import Link from 'next/link';
 import styles from '@/styles/Card.module.scss';
-import { Tag, TagContainer } from '@/components/blog/Tag';
+import S from './styles';
+import { Tag } from '@/components/blog/Tag/Tag';
 
 export default function Card({
   card: { image, altTag, title, content, link, linkText, tagList },
@@ -30,11 +31,11 @@ export default function Card({
         {title}
       </h2>
       {tagList && tagList.length > 0 ? (
-        <TagContainer>
+        <S.TagContainer>
           {tagList.slice(0, 8).map((tag, i) => (
             <Tag key={i} text={tag} />
           ))}
-        </TagContainer>
+        </S.TagContainer>
       ) : null}
       <div className={styles.content}>
         <p>
